requirements.txt
--extra-index-url https://pypi.fury.io/reW6hdjFCrSbpo3ehrBu/alpha-i/

alphai_delphi>=1.1.1,<2.0.0
alphai_finance==2.0.0
alphai-time-series>=0.0.4
alphai-alcova-datasource>=0.0.1,<1.0.0
alphai-data-sources>=1.0.1
alphai_feature_generation>=1.2.1,<2.0.0
alphai_covariance>=0.1.4
tensorflow>=1.4.0
numpy>=1.12.0
pandas==0.18.1
pandas-market-calendars>=0.12
scikit-learn>=0.19.0
pyaml

